+++ CHANGELOG.md
@@ -10,9 +10,15 @@ Items marked with :star: is the major change why we release a new version.
 * **BREAKING**
 Replace ` Line  CODE  JT   JF      K` with ` Label  CODE  JT   JF      K`.
 Replace disasm output ` 0001: 0x20 0x00 0x00 0x00000004 $A = $arch` with ` L0001: 0x20 0x00 0x00 0x00000004 $A = $arch`. The good news we add support for label, we can use `goto Label_Allow`, also something like `Label_Allow:` in the start of a line as Label Decalarion.
+* :star: Add seize subfunction in trace pid :link: [#23]
 * :star: Remove `-mno-omit-leaf-frame-pointer` flag if not supported by cc
-* :star: Fix asm output endianness problem
+* :star: Fix asm and disasm endianness problems :link: [#22]
+* :star: Handle '>=' '>' '<=' '<' correctly :link: [#25]
 * Print messages in dynamic tests
+
+[#22]: https://github.com/dbgbgtf1/Ceccomp/issues/22
+[#25]: https://github.com/dbgbgtf1/Ceccomp/issues/25
+[#23]: https://github.com/dbgbgtf1/Ceccomp/issues/23
 ## 3.5
 * **BREAKING** :star: Now `trace` and `probe` no more accepts `-a`; arch is automatically retrieved via ptrace
 * :star: Add `i686` arch support